funcionalidades Aplicación Autenticación de Usuario(firebase) Gestión de Productos (Crear, Leer, Actualizar, Eliminar) Gestión de Pedidos (gestionamos los valores negativo positivos y lo mostramos en pantalla) Operaciones CRUD SQLite(matemáticas en la base de datos) Diseño y Usabilidad (super agil y liviana) Documentación y Comentarios (archivo y video de presentación) Exportación de base de datos .db (carpeta Descargas) Importación de base de datos .db (carpeta Descargas del movil) Exportación de base de datos a Excel Activity Perfil de usuario con (firebase personalizado) Buscador de Items dentro de la base de datos menu diálogo layou egreso total layou ingreso total creacion de codigo de acceso personalizado para que no use login agregar items personalizados para usar de forma recurrente contabilidad donde el usuiario puede ver por dia mes año su contabilidad corregido edicion negativo informa al usario

+++ app/src/main/java/com/example/tiendacontrol/model/Items.java
@@ -25,8 +25,8 @@ public class Items {
         this.producto = producto;
     }
 
-    public double getValor() {
-        return valor;
+    public int getValor() {
+        return (int) valor;
     }
 
     public void setValor(double valor) {
